all you need to compile
this is the minimum amount of work needed to do to get a SDR# plugin to
compile.  The original is ZoomFFT, and I hacked it up a little and got
waveletPWR.  While it doesn't actually do any wavelet processing, I've
shown what you have to do to make small edits in both the control panel
and spectrum analyzer output windows.  just do a search for EDIT to find
a description of changes I made, and search for TODO to find things that
require changing, but I hadn't changed yet.